opengl: the engine will select default anisotropy filtering level
  on some modern GPUs anisotropy works even without mipmaps and
  bilinear filtering, and makes the image much better on higher
  resolutions. it is totally worth several FPS. so now with the
  fresh config the engine will select a good level for you.

FossilOrigin-Name: 3fc2249b990e937fcbc1becb7127414706cf7aab5f3e26f71bb2484db1d406f0

+++ source/render/hwgl/gl_local.h
@@ -984,7 +984,7 @@ protected:
   vuint8 atlases_updated[NUM_BLOCK_SURFS];
 
   GLenum ClampToEdge;
-  GLfloat max_anisotropy; // 1.0: off
+  GLint max_anisotropy; // 1: off
   bool anisotropyExists;
 
   bool usingFPZBuffer;
